create order, config role, update order state, add product, get all order

+++ src/main/java/com/example/demo/entity/Image.java
@@ -12,10 +12,7 @@ import lombok.Setter;
 public class Image {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String url;
     private String publicId;
-
-
 }
